+++ utils/errors.go
@@ -105,3 +105,8 @@ func WrapMoveToFileError(err error) error {
 	}
 	return fmt.Errorf("moveToFile error: %w", err)
 }
+
+// WrapStatError returns a wrapped Stat error
+func WrapStatError(err error) error {
+	return fmt.Errorf("stat error: %w", err)
+}