+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Update vfs to implement io/fs's fs.FS interface fixes #163
 
 ## [v7.5.1] - 2025-07-24
 ### Fixed